requirements.txt
# =============================================================================
# CareerLens - AI Career Intelligence Platform Dependencies
# =============================================================================
# Optimized for Streamlit Cloud deployment
# All libraries listed are actively used in the codebase

# -----------------------------------------------------------------------------
# Core Framework & Web
# -----------------------------------------------------------------------------
streamlit>=1.28.0,<2.0.0
requests>=2.31.0,<3.0.0

# -----------------------------------------------------------------------------
# Data Processing & ML
# -----------------------------------------------------------------------------
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0      # Used for cosine_similarity
Pillow>=10.0.0                  # Image processing

# -----------------------------------------------------------------------------
# Document Processing (Resume Upload/Export)
# -----------------------------------------------------------------------------
PyPDF2>=3.0.0,<4.0.0            # PDF text extraction
python-docx>=1.0.0,<2.0.0       # DOCX text extraction & generation

# -----------------------------------------------------------------------------
# Vector Store & Semantic Search
# -----------------------------------------------------------------------------
pinecone-client>=3.0.0,<4.0.0   # Pinecone vector database
sentence-transformers>=2.2.0,<3.0.0  # Embedding model (all-MiniLM-L6-v2)

# Optional: ChromaDB for local vector store (CareerLens)
chromadb>=0.4.0,<0.6.0

# -----------------------------------------------------------------------------
# AI/LLM APIs
# -----------------------------------------------------------------------------
openai>=1.0.0,<2.0.0            # Azure OpenAI API
tiktoken>=0.5.0,<1.0.0          # Token counting for API rate limiting

# -----------------------------------------------------------------------------
# Data Visualization
# -----------------------------------------------------------------------------
matplotlib>=3.7.0,<4.0.0        # Charts and graphs
plotly>=5.15.0,<6.0.0           # Interactive visualizations

# -----------------------------------------------------------------------------
# PDF Generation (Resume Export)
# -----------------------------------------------------------------------------
reportlab>=4.0.0,<5.0.0         # Professional PDF resume generation

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
python-dateutil>=2.8.0,<3.0.0   # Date handling
httpx
